+++ tests/js/client/server_parameters/test-javascript-transactions-off-allow-admin-execute.js
@@ -30,7 +30,7 @@
 
 if (getOptions === true) {
   return {
-    'javascript.tasks': "false",
+    'javascript.transactions': "false",
     'javascript.allow-admin-execute': "true"
   };
 }
@@ -41,20 +41,29 @@ const db = require('internal').db;
 const FoxxManager = require('@arangodb/foxx/manager');
 const path = require('path');
 const internal = require('internal');
-const basePath = path.resolve(internal.pathForTesting('common'), 'test-data', 'apps', 'execute-task');
+const basePath = path.resolve(internal.pathForTesting('common'), 'test-data', 'apps', 'execute-transaction');
 
 require("@arangodb/test-helper").waitForFoxxInitialized();
 
 function testSuite() {
+  const cn = "UnitTestsTransaction";
+
   return {
-    testJavaScriptTask : function() {
-      // JavaScript tasks should not work
+    setUp : function() {
+      db._drop(cn);
+      db._create(cn);
+    },
+
+    tearDown : function() {
+      db._drop(cn);
+    },
+
+    testJavaScriptTransaction : function() {
+      // JavaScript transactions should be affected by the setting and fail
       try {
-        require('@arangodb/tasks').register({
-          command: function() {
-            require("console").log("this task must not run!");
-            throw "peng!";
-          }
+        db._executeTransaction({ 
+          collections: { read: cn },
+          action: function() {},
         });
         fail();
       } catch (err) {
@@ -62,7 +71,7 @@ function testSuite() {
       }
     },
     
-    testJavaScriptTaskFromFoxx : function() {
+    testJavaScriptTransactionFromFoxx : function() {
       const mount = '/test';
 
       FoxxManager.install(basePath, mount);
@@ -73,6 +82,37 @@ function testSuite() {
       } finally {
         FoxxManager.uninstall(mount, {force: true});
       } 
+    },
+    
+    testJavaScriptTransactionViaAdminExecute : function() {
+      let body = `require('@arangodb').db._executeTransaction({ collections: { read: "${cn}" }, action: function() {} }); return "ok!"; `;
+
+      let res = arango.POST('/_db/_system/_admin/execute?returnBodyAsJSON=true', body);
+      assertEqual("ok!", res);
+    },
+
+    testNonJavaScriptTransaction : function() {
+      // non-JavaScript transactions should not be affected by the setting
+      const opts = {
+        collections: { write: cn },
+      };
+      
+      const trx = db._createTransaction(opts);
+      try {
+        const tc = trx.collection(cn);
+        assertEqual(0, tc.count());
+        for (let i = 0; i < 10; ++i) {
+          tc.insert({ _key: "test" + i });
+        }
+        assertEqual(10, tc.count());
+        assertEqual(0, db[cn].count());
+        trx.commit();
+      
+        assertEqual(10, db[cn].count());
+      } catch (err) {
+        trx.abort();
+        throw err;
+      }
     },
 
   };
